Project directory tree, main architecture, elementary functionality.
Most of the information regarding suggestions is provided inside the codebase (which is currently very easy to follow due to its small size).
I am not certain if I have written the `.gitgnore` file. That being said, it may be edited in the future, especially if we adjust the codebase for the ELSED line detection algorithm, in which case we won't wish to commit the built library (only the source files).